+++ src/test/java/racinggame/domain/race/RaceTest.java
@@ -1,10 +1,9 @@
-package racinggame;
+package racinggame.domain.race;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import racinggame.domain.strategy.MoveStrategy;
-import racinggame.domain.race.Race;
 import racinggame.domain.strategy.RandomMoveStrategy;
 
 import static org.assertj.core.api.Assertions.assertThat;
