+++ src/test/java/racinggame/domain/racingcar/RacingCarCreationTest.java
@@ -1,9 +1,8 @@
-package racinggame;
+package racinggame.domain.racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import racinggame.domain.strategy.MoveStrategy;
-import racinggame.domain.racingcar.RacingCars;
 
 import java.util.Arrays;
 import java.util.List;
