+++ src/test/java/racinggame/view/RacingGameInputViewTest.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.view;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
@@ -7,7 +7,6 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
-import racinggame.view.ConsoleInputView;
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
